// src/modules/booking/booking.constant.d.ts
export declare const BookingMessages: {
    CREATED: string;
    UPDATED: string;
    NOT_FOUND: string;
    ALREADY_EXISTS: string;
    INVALID_STATUS: string;
    NOT_AUTHORIZED: string;
    TECHNICIAN_NOT_AVAILABLE: string;
    TECHNICIAN_NOT_APPROVED: string;
    SERVICE_NOT_FOUND: string;
    CANNOT_CANCEL: string;
};
//# sourceMappingURL=booking.constant.d.ts.map